refactored forms and social login
