+++ src/common/PrivateRoute.js
@@ -23,4 +23,4 @@ const PrivateRoute = ({ component: Component, authenticated, ...rest }) => (
     />
 );
   
-export default PrivateRoute
+export default PrivateRoute;
