+++ src/components/Forms/SignupForm/SignupForm.js
@@ -1,0 +1,68 @@
+import React, { Component } from 'react';
+
+//Components to Import
+import { signup } from '../../../util/APIUtils';
+import Alert from 'react-s-alert';
+
+class SignupForm extends Component {
+    constructor(props) {
+        super(props);
+        this.state = {
+            name: '',
+            email: '',
+            password: ''
+        }
+    }
+
+    handleInputChange = (event) => {
+        const target = event.target;
+        const inputName = target.name;        
+        const inputValue = target.value;
+
+        this.setState({
+            [inputName] : inputValue
+        });        
+    }
+
+    handleSubmit = (event) => {
+        event.preventDefault();   
+
+        const signUpRequest = Object.assign({}, this.state);
+
+        signup(signUpRequest)
+        .then(response => {
+            Alert.success("You're successfully registered. Please login to continue!");
+            this.props.history.push("/login");
+        }).catch(error => {
+            Alert.error((error && error.message) || 'Oops! Something went wrong. Please try again!');            
+        });
+    }
+
+    render() {
+        return (
+            <form onSubmit={this.handleSubmit}>
+                <div className="form-item">
+                    <input type="text" name="name" 
+                        className="form-control" placeholder="Name"
+                        value={this.state.name} onChange={this.handleInputChange} required/>
+                </div>
+                <div className="form-item">
+                    <input type="email" name="email" 
+                        className="form-control" placeholder="Email"
+                        value={this.state.email} onChange={this.handleInputChange} required/>
+                </div>
+                <div className="form-item">
+                    <input type="password" name="password" 
+                        className="form-control" placeholder="Password"
+                        value={this.state.password} onChange={this.handleInputChange} required/>
+                </div>
+                <div className="form-item">
+                    <button type="submit" className="btn btn-block btn-primary" >Sign Up</button>
+                </div>
+            </form>                    
+
+        );
+    }
+}
+
+export default SignupForm;
