+++ src/user/login/Login.js
@@ -1,14 +1,17 @@
 import React, { Component } from 'react';
 import './Login.css';
-import { GOOGLE_AUTH_URL, FACEBOOK_AUTH_URL, GITHUB_AUTH_URL, ACCESS_TOKEN } from '../../constants';
-import { login } from '../../util/APIUtils';
-import { Link, Redirect } from 'react-router-dom'
-import fbLogo from '../../img/fb-logo.png';
-import googleLogo from '../../img/google-logo.png';
-import githubLogo from '../../img/github-logo.png';
+import { Link, Redirect } from 'react-router-dom';
+
+//Components to Import
 import Alert from 'react-s-alert';
+import LoginForm from '../../components/Forms/LoginForm/LoginForm';
+import SocialLoginForm from '../../components/Forms/SocialLoginForm/SocialLoginForm';
 
 class Login extends Component {
+    state = {
+        isLoginOrSignup: "Login"
+    }
+
     componentDidMount() {
         // If the OAuth2 login encounters an error, the user is redirected to the /login page with an error.
         // Here we display the error and then remove the error query parameter from the location.
@@ -38,7 +41,7 @@ class Login extends Component {
             <div className="login-container">
                 <div className="login-content">
                     <h1 className="login-title">Login to SpringSocial</h1>
-                    <SocialLogin />
+                    <SocialLoginForm displayText={this.state.isLoginOrSignup}/>
                     <div className="or-separator">
                         <span className="or-text">OR</span>
                     </div>
@@ -46,79 +49,6 @@ class Login extends Component {
                     <span className="signup-link">New user? <Link to="/signup">Sign up!</Link></span>
                 </div>
             </div>
-        );
-    }
-}
-
-class SocialLogin extends Component {
-    render() {
-        return (
-            <div className="social-login">
-                <a className="btn btn-block social-btn google" href={GOOGLE_AUTH_URL}>
-                    <img src={googleLogo} alt="Google" /> Log in with Google</a>
-                <a className="btn btn-block social-btn facebook" href={FACEBOOK_AUTH_URL}>
-                    <img src={fbLogo} alt="Facebook" /> Log in with Facebook</a>
-                <a className="btn btn-block social-btn github" href={GITHUB_AUTH_URL}>
-                    <img src={githubLogo} alt="Github" /> Log in with Github</a>
-            </div>
-        );
-    }
-}
-
-
-class LoginForm extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            email: '',
-            password: ''
-        };
-        this.handleInputChange = this.handleInputChange.bind(this);
-        this.handleSubmit = this.handleSubmit.bind(this);
-    }
-
-    handleInputChange(event) {
-        const target = event.target;
-        const inputName = target.name;        
-        const inputValue = target.value;
-
-        this.setState({
-            [inputName] : inputValue
-        });        
-    }
-
-    handleSubmit(event) {
-        event.preventDefault();   
-
-        const loginRequest = Object.assign({}, this.state);
-
-        login(loginRequest)
-        .then(response => {
-            sessionStorage.setItem(ACCESS_TOKEN, response.accessToken);
-            Alert.success("You're successfully logged in!");
-            this.props.history.push("/");
-        }).catch(error => {
-            Alert.error((error && error.message) || 'Oops! Something went wrong. Please try again!');
-        });
-    }
-    
-    render() {
-        return (
-            <form onSubmit={this.handleSubmit}>
-                <div className="form-item">
-                    <input type="email" name="email" 
-                        className="form-control" placeholder="Email"
-                        value={this.state.email} onChange={this.handleInputChange} required/>
-                </div>
-                <div className="form-item">
-                    <input type="password" name="password" 
-                        className="form-control" placeholder="Password"
-                        value={this.state.password} onChange={this.handleInputChange} required/>
-                </div>
-                <div className="form-item">
-                    <button type="submit" className="btn btn-block btn-primary">Login</button>
-                </div>
-            </form>                    
         );
     }
 }
